refactor(architecture): reorganize project structure to clean architecture layers
