+++ src/Application/Tracking/Command/TrackVisitCommand.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yomali\Tracker\Application\Tracking\Command;
+
+/**
+ * Command to track a new visit
+ */
+final readonly class TrackVisitCommand
+{
+    public function __construct(
+        public string $ipAddress,
+        public string $pageUrl,
+    ) {
+    }
+}
