--- src/Tracking/Application/Command/TrackVisitCommand.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yomali\Tracker\Tracking\Application\Command;
-
-/**
- * Command to track a new visit
- */
-final readonly class TrackVisitCommand
-{
-    public function __construct(
-        public string $ipAddress,
-        public string $pageUrl,
-    ) {
-    }
-}
