--- tests/Unit/Tracking/Infrastructure/Persistence/Connection/MySQLConnectionTest.php
@@ -1,222 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yomali\Tracker\Tests\Unit\Tracking\Infrastructure\Persistence\Connection;
-
-use Yomali\Tracker\Tests\Unit\UnitTestCase;
-use Yomali\Tracker\Tracking\Infrastructure\Persistence\Connection\MySQLConnection;
-
-/**
- * @group unit
- */
-final class MySQLConnectionTest extends UnitTestCase
-{
-    protected function setUp(): void
-    {
-        parent::setUp();
-        
-        // Set up test database configuration from environment
-        $this->setTestDbConfig();
-    }
-
-    protected function tearDown(): void
-    {
-        // Reset singleton after each test - do this AFTER running test
-        parent::tearDown();
-        MySQLConnection::reset();
-    }
-
-    public function testSingletonReturnsSameInstance(): void
-    {
-        $instance1 = MySQLConnection::getInstance();
-        $instance2 = MySQLConnection::getInstance();
-
-        $this->assertSame($instance1, $instance2);
-    }
-
-    public function testGetPdoReturnsPdoInstance(): void
-    {
-        $connection = MySQLConnection::getInstance();
-        $pdo = $connection->getPdo();
-
-        $this->assertInstanceOf(\PDO::class, $pdo);
-    }
-
-    public function testResetClearsInstance(): void
-    {
-        $instance1 = MySQLConnection::getInstance();
-        MySQLConnection::reset();
-        $instance2 = MySQLConnection::getInstance();
-
-        $this->assertNotSame($instance1, $instance2);
-    }
-
-    public function testSingletonPatternIsImplementedCorrectly(): void
-    {
-        $reflection = new \ReflectionClass(MySQLConnection::class);
-
-        // Verify __clone is private
-        $cloneMethod = $reflection->getMethod('__clone');
-        $this->assertTrue($cloneMethod->isPrivate());
-
-        // Verify constructor is private
-        $constructor = $reflection->getConstructor();
-        $this->assertTrue($constructor->isPrivate());
-
-        // Verify singleton instance property exists and is private
-        $instanceProperty = $reflection->getProperty('instance');
-        $this->assertTrue($instanceProperty->isPrivate());
-        $this->assertTrue($instanceProperty->isStatic());
-    }
-
-    public function testGetRequiredEnvReturnsValue(): void
-    {
-        $_ENV['TEST_VAR'] = 'test_value';
-
-        $connection = MySQLConnection::getInstance();
-        $result = $this->callPrivateMethod($connection, 'getRequiredEnv', 'TEST_VAR');
-
-        $this->assertEquals('test_value', $result);
-
-        // Clean up
-        unset($_ENV['TEST_VAR']);
-    }
-
-    public function testGetRequiredEnvThrowsWhenVariableNotSet(): void
-    {
-        $this->expectException(\RuntimeException::class);
-        $this->expectExceptionMessage("Required environment variable 'NONEXISTENT_VAR' is not set");
-
-        $connection = MySQLConnection::getInstance();
-        $this->callPrivateMethod($connection, 'getRequiredEnv', 'NONEXISTENT_VAR');
-    }
-
-    public function testGetRequiredEnvThrowsWhenVariableEmpty(): void
-    {
-        $_ENV['EMPTY_VAR'] = '';
-
-        $this->expectException(\RuntimeException::class);
-        $this->expectExceptionMessage("Required environment variable 'EMPTY_VAR' is not set");
-
-        $connection = MySQLConnection::getInstance();
-        $this->callPrivateMethod($connection, 'getRequiredEnv', 'EMPTY_VAR');
-
-        // Clean up
-        unset($_ENV['EMPTY_VAR']);
-    }
-
-    public function testGetRequiredEnvUsesGetenvAsFallback(): void
-    {
-        // Remove from $_ENV but set via putenv (simulates Docker environment)
-        unset($_ENV['FALLBACK_VAR']);
-        putenv('FALLBACK_VAR=fallback_value');
-
-        $connection = MySQLConnection::getInstance();
-        $result = $this->callPrivateMethod($connection, 'getRequiredEnv', 'FALLBACK_VAR');
-
-        $this->assertEquals('fallback_value', $result);
-
-        // Clean up
-        putenv('FALLBACK_VAR');
-    }
-
-    public function testLoadEnvironmentHandlesFileExists(): void
-    {
-        $connection = MySQLConnection::getInstance();
-        
-        // Call loadEnvironment - should not throw exceptions if .env file exists
-        $this->callPrivateMethod($connection, 'loadEnvironment');
-        
-        $this->assertTrue(true); // If we reach here, no exceptions were thrown
-    }
-
-
-    public function testConstructorThrowsOnPdoConnectionFailure(): void
-    {
-        // Reset singleton FIRST to avoid using cached instance
-        MySQLConnection::reset();
-        
-        // Temporarily rename .env file to prevent it from overriding our invalid values
-        $envFile = '/var/www/.env';
-        $envBackup = '/var/www/.env.backup.test';
-        $envMoved = false;
-        
-        if (file_exists($envFile)) {
-            rename($envFile, $envBackup);
-            $envMoved = true;
-        }
-        
-        // Backup original environment
-        $originalVars = [
-            'DB_HOST' => $_ENV['DB_HOST'] ?? null,
-            'DB_NAME' => $_ENV['DB_NAME'] ?? null,
-            'DB_USER' => $_ENV['DB_USER'] ?? null,
-            'DB_PASSWORD' => $_ENV['DB_PASSWORD'] ?? null,
-            'DB_PORT' => $_ENV['DB_PORT'] ?? null,
-        ];
-
-        try {
-            // Set up environment with invalid values that will cause immediate PDO failure
-            $_ENV['DB_HOST'] = 'invalid_host_999.999.999.999';
-            $_ENV['DB_NAME'] = 'invalid_db';  
-            $_ENV['DB_USER'] = 'invalid_user';
-            $_ENV['DB_PASSWORD'] = 'invalid_password';
-            $_ENV['DB_PORT'] = 'invalid_port'; // Invalid port will cause immediate DSN error
-
-            // Also clear putenv to ensure invalid values are used
-            putenv('DB_HOST=invalid_host_999.999.999.999');
-            putenv('DB_NAME=invalid_db');
-            putenv('DB_USER=invalid_user');
-            putenv('DB_PASSWORD=invalid_password');
-            putenv('DB_PORT=invalid_port');
-
-            $this->expectException(\RuntimeException::class);
-            $this->expectExceptionMessage('Database connection failed:');
-            
-            // This should trigger constructor's PDO exception handling
-            MySQLConnection::getInstance();
-            
-        } finally {
-            // Always restore original environment variables
-            foreach ($originalVars as $key => $value) {
-                if ($value !== null) {
-                    $_ENV[$key] = $value;
-                    putenv("{$key}={$value}");
-                } else {
-                    unset($_ENV[$key]);
-                }
-            }
-            
-            // Restore .env file if we moved it
-            if ($envMoved && file_exists($envBackup)) {
-                rename($envBackup, $envFile);
-            }
-            
-            // Reset singleton to clean state
-            MySQLConnection::reset();
-        }
-    }
-
-    public function testCloneMethodPreventsCloning(): void
-    {
-        $reflection = new \ReflectionClass(MySQLConnection::class);
-        $cloneMethod = $reflection->getMethod('__clone');
-        
-        // Verify the method is private (prevents external cloning)
-        $this->assertTrue($cloneMethod->isPrivate(), '__clone method should be private');
-        
-        // Test that the method exists and is callable via reflection
-        // This covers the method even though it has an empty body
-        $connection = MySQLConnection::getInstance();
-        
-        // Set method accessible and invoke it to cover the lines
-        $cloneMethod->setAccessible(true);
-        $cloneMethod->invoke($connection);
-        
-        // If we reach here, the method was successfully called
-        // The empty body is now covered
-        $this->assertTrue(true);
-    }
-
-}
